refactor(config): extract nested tests module
Hoist the ~2400-line #[cfg(test)] mod tests body into config/tests.rs;
mod.rs becomes a curated index (~160 lines) with explicit test re-exports
for Raw types and load helpers.
